# contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_knowledge_documents_post": {
        "properties": {
          "classification": {
            "default": "internal",
            "title": "Classification",
            "type": "string"
          },
          "domain": {
            "default": "shared",
            "title": "Domain",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "scope": {
            "default": "tenant",
            "title": "Scope",
            "type": "string"
          },
          "source_version": {
            "default": "1",
            "title": "Source Version",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "file",
          "title"
        ],
        "title": "Body_upload_document_api_v1_knowledge_documents_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestJobView": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "chunk_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Count"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "job_id",
          "status",
          "title",
          "filename",
          "scope",
          "attempts",
          "error",
          "document_id",
          "chunk_count",
          "created_at",
          "updated_at"
        ],
        "title": "IngestJobView",
        "type": "object"
      },
...
      "KnowledgeDocumentView": {
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "index_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index Version"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "source_version": {
            "title": "Source Version",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "title",
          "domain",
          "classification",
          "source_version",
          "index_version",
          "chunk_count",
          "created_at",
          "scope"
        ],
        "title": "KnowledgeDocumentView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Digital Worker Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/documents": {
      "get": {
        "operationId": "list_documents_api_v1_knowledge_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeDocumentView"
                  },
                  "title": "Response List Documents Api V1 Knowledge Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Documents",
        "tags": [
          "knowledge"
        ]
      },
      "post": {
        "operationId": "upload_document_api_v1_knowledge_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_knowledge_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Document",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/documents/jobs/{job_id}": {
      "get": {
        "operationId": "get_ingest_job_api_v1_knowledge_documents_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ingest Job",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document_api_v1_knowledge_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Document",
        "tags": [
          "knowledge"
        ]
      }
    },
...
  }
}